requirements.txt
-e git+https://github.com/darklow/django-suit@v2#egg=django-suit
django-auth-ldap==2.1.*
django-extensions==2.2.*
django==2.2.*
djangorestframework==3.11.*
fabric==2.5.*
gunicorn==20.0.*
ipaddress==1.0.*
mysqlclient==1.4.6
sentry-sdk==0.13.*
